@@ -4,7 +4,7 @@ namespace DntSite.Web.Features.Courses.Models;
 
 public class CourseQuestionsModel
 {
-    public IList<CourseQuestion> CourseQuestions { set; get; } = new List<CourseQuestion>();
+    public IList<CourseQuestion> CourseQuestions { set; get; } = [];
 
     public Course? Course { set; get; }
 }
