@@ -6,7 +6,7 @@ public class CourseSideMenuModel
 {
     public Course? ThisCourse { set; get; }
 
-    public IList<CourseQuestionComment> CourseQuestionComments { set; get; } = new List<CourseQuestionComment>();
+    public IList<CourseQuestionComment> CourseQuestionComments { set; get; } = [];
 
-    public IList<CourseTopicComment> CourseTopicComments { set; get; } = new List<CourseTopicComment>();
+    public IList<CourseTopicComment> CourseTopicComments { set; get; } = [];
 }
