@@ -16,15 +16,15 @@ public abstract class BaseInteractiveEntity<TSelfEntity, TVisitorEntity, TBookma
     where TCommentsBookmarkEntity : BaseBookmarkEntity<TCommentsEntity>
     where TCommentsReactionEntity : BaseReactionEntity<TCommentsEntity>
 {
-    public virtual ICollection<TCommentsEntity> Comments { set; get; } = new List<TCommentsEntity>();
+    public virtual ICollection<TCommentsEntity> Comments { set; get; } = [];
 
-    public virtual ICollection<TTagEntity> Tags { set; get; } = new List<TTagEntity>();
+    public virtual ICollection<TTagEntity> Tags { set; get; } = [];
 
-    public virtual ICollection<TReactionEntity> Reactions { set; get; } = new List<TReactionEntity>();
+    public virtual ICollection<TReactionEntity> Reactions { set; get; } = [];
 
-    public virtual ICollection<TBookmarkEntity> Bookmarks { set; get; } = new List<TBookmarkEntity>();
+    public virtual ICollection<TBookmarkEntity> Bookmarks { set; get; } = [];
 
-    public virtual IEnumerable<TVisitorEntity> Visitors { set; get; } = new List<TVisitorEntity>();
+    public virtual IEnumerable<TVisitorEntity> Visitors { set; get; } = [];
 
-    public virtual IEnumerable<TUserFileEntity> UserFiles { set; get; } = new List<TUserFileEntity>();
+    public virtual IEnumerable<TUserFileEntity> UserFiles { set; get; } = [];
 }
