@@ -4,9 +4,9 @@ namespace DntSite.Web.Features.Stats.Models;
 
 public class AnnualStatisticsModel
 {
-    public IList<AnnualStatisticsInfo> AnnualStatisticsInfo { set; get; } = new List<AnnualStatisticsInfo>();
+    public IList<AnnualStatisticsInfo> AnnualStatisticsInfo { set; get; } = [];
 
     public int PersianYear { set; get; }
 
-    public IList<SelectListItem> PersianYears { set; get; } = new List<SelectListItem>();
+    public IList<SelectListItem> PersianYears { set; get; } = [];
 }
