@@ -6,9 +6,9 @@ public class ShowReactionsUsersListModel
 {
     public ShowReactionsModel ReactionsInfo { set; get; } = new();
 
-    public IList<User?> ThumbsDownUsers { set; get; } = new List<User?>();
+    public IList<User?> ThumbsDownUsers { set; get; } = [];
 
-    public IList<User?> ThumbsUpUsers { set; get; } = new List<User?>();
+    public IList<User?> ThumbsUpUsers { set; get; } = [];
 
     /// <summary>
     ///     This will allows you to convert this object to a tuple.
