@@ -13,7 +13,7 @@ public partial class ServerInfo
 
     [Inject] internal IKeyManager KeyManager { set; get; } = null!;
 
-    private List<IKey> GetKeysList() => KeyManager.GetAllKeys().OrderByDescending(key => key.CreationDate).ToList();
+    private List<IKey> GetKeysList() => [..KeyManager.GetAllKeys().OrderByDescending(key => key.CreationDate)];
 
     protected override async Task OnInitializedAsync()
     {
