@@ -189,13 +189,13 @@ public class DailyNewsItemsService(
 
         if (lastRecord is null)
         {
-            return new List<DailyNewsItem>();
+            return [];
         }
 
         var year = int.Parse(lastRecord.Audit.CreatedAtPersian.AsSpan(start: 0, length: 4),
             CultureInfo.InvariantCulture);
 
-        year = year - pageNumber;
+        year -= pageNumber;
         var fromDate = string.Create(CultureInfo.InvariantCulture, $"{year}/01/01");
         var toDate = string.Create(CultureInfo.InvariantCulture, $"{year}/12/30");
 
